saves  timer
it saves each time you clock in, so our data is safe. but we need to handle crash and refresh. so on refresh, or crash or something save the data. retrieve it on startup and continue the useeffect automatically. maybe set a state that you save to localstorage and run the stopwatch based on that